refactor: 테스트 클래스에서 'Impl' 제거
